+++ streams/src/test/java/org/apache/kafka/streams/StreamsConfigTest.java
@@ -31,6 +31,7 @@ import org.apache.kafka.common.serialization.Serdes;
 import org.apache.kafka.common.serialization.Serializer;
 import org.apache.kafka.common.utils.Utils;
 import org.apache.kafka.streams.errors.StreamsException;
+import org.apache.kafka.streams.internals.UpgradeFromValues;
 import org.apache.kafka.streams.processor.FailOnInvalidTimestamp;
 import org.apache.kafka.streams.processor.TimestampExtractor;
 import org.apache.kafka.streams.processor.internals.DefaultKafkaClientSupplier;
@@ -1392,7 +1393,7 @@ public class StreamsConfigTest {
 
     @Test
     public void shouldSupportAllUpgradeFromValues() {
-        for (final StreamsConfig.UpgradeFromValues upgradeFrom : StreamsConfig.UpgradeFromValues.values()) {
+        for (final UpgradeFromValues upgradeFrom : UpgradeFromValues.values()) {
             props.put(StreamsConfig.UPGRADE_FROM_CONFIG, upgradeFrom.toString());
             try {
                 new StreamsConfig(props);
